Use a var to get docker flags -i and -t on web-dev
It's actually really hard to figure out if we run on a tty and if we can
consume interactive commands because we run in the foreground.

Especially it is hard if we try to wrap all this in a single command.

So we offer a environment-variable instead "WEB_DEV_IT" to let the user
decide in which environment the command will run.